+++ src/cufinder/services/__init__.py
@@ -21,6 +21,14 @@ from .clo import Clo
 from .cse import Cse
 from .pse import Pse
 from .lcuf import Lcuf
+from .bcd import Bcd
+from .ccp import Ccp
+from .isc import Isc
+from .cbc import Cbc
+from .csc import Csc
+from .csn import Csn
+from .nao import Nao
+from .naa import Naa
 
 __all__ = [
     "BaseService",
@@ -44,4 +52,12 @@ __all__ = [
     "Cse",
     "Pse",
     "Lcuf",
+    "Bcd",
+    "Ccp",
+    "Isc",
+    "Cbc",
+    "Csc",
+    "Csn",
+    "Nao",
+    "Naa",
 ]